refactor: 바뀐 types 폴더로 새로 Import
