+++ src/component/CardPeriod/CardPeriodInput.tsx
@@ -9,12 +9,15 @@ import {
 } from '../../styles/@common/text/text.style';
 
 import { cardPeriodInputLayout } from './CardPeriodInput.style';
-import { CardExpirationDate, CardExpirationDateError } from '../../hooks';
 import { CARD_EXPIRATION_ERROR, CARD_EXPIRATION } from '../../constants';
 import {
   inputContainer,
   inputSection,
 } from '../../styles/@common/inputContainer.style';
+import {
+  CardExpirationDate,
+  CardExpirationDateError,
+} from '../../../types/types';
 
 type CardPeriodInputProps = {
   cardExpirationDate: CardExpirationDate;
